Fix buggy anchoring in landscape orientation - 2

+++ FantasmoSDK/Classes/Utilities/MovingAverage.swift
@@ -9,19 +9,19 @@ class MovingAverage {
     
     private var index = 0
     private let period: Int
-    private var samples: Array<Double>
+    private var samples: Array<Float>
     
     init(period: Int = 30) {
         self.period = period
         samples = []
     }
     
-    var average: Double {
-        let sum = samples.reduce(0.0, +)
-        return sum / Double(samples.count)
+    var average: Float {
+        let sum = samples.reduce(0, +)
+        return sum / Float(samples.count)
     }
     
-    func addSample(value: Double) -> Double {
+    func addSample(value: Float) -> Float {
         if samples.count == period {
             samples[index] = value
             index = (index + 1) % period
